requirements.txt
akshare==1.11.18
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
APScheduler==3.10.4
pandas==2.0.3
plotly==5.18.0
dash==2.14.2
requests==2.31.0
python-dotenv==1.0.0 